+++ ch6/sequence/test-NodeSequence.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+g++ test-NodeSequence.cpp -o test-NodeSequence
+./test-NodeSequence
+
+rm test-NodeSequence